Changed Player mechanics/skybox overhaul
made a number of changes to the player movement mechanics and wall running which now only wall runs on objects "tagged" as "climbable" that way the player doesn't wall run on random crap. Also fixed the door opening so it's smooth. Fixed the skybox as well to give it that outer space feel and removed the sword mechanics